+++ sites/plus.skeleton.dev/src/lib/auth/supported-oauth-providers.ts
