+++ sites/plus.skeleton.dev/src/lib/components/auth/sign-in-button.svelte
@@ -7,8 +7,8 @@
 </script>
 
 <script lang="ts">
+	import type { SupportedOAuthProvider } from '$lib/auth/supported-oauth-providers';
 	import { signIn } from '$lib/remote/auth/sign-in.remote';
-	import type { SupportedOAuthProvider } from '$lib/server/auth/supported-oauth-providers';
 
 	const { children, oAuthProvider, ...attributes }: Props = $props();
 </script>
