+++ sites/plus.skeleton.dev/src/lib/server/auth/auth.ts
@@ -1,16 +1,16 @@
 import { betterAuth } from 'better-auth';
 import { drizzleAdapter } from 'better-auth/adapters/drizzle';
-import { database } from '$lib/server/database/database';
+import { db } from '$lib/server/database/db';
 import { env } from '$env/dynamic/private';
-import * as schema from '$lib/server/database/schema/schema';
+import * as schema from '$lib/server/database/schema';
 import { getRequestEvent } from '$app/server';
 import { sveltekitCookies } from 'better-auth/svelte-kit';
-import type { SupportedOAuthProvider } from './supported-oauth-providers';
-import { getOrigin } from '$lib/app/get-origin';
+import type { SupportedOAuthProvider } from '$lib/auth/supported-oauth-providers';
+import { getAppOrigin } from '$lib/utility/get-app-origin';
 
 export const auth = betterAuth({
-	baseURL: getOrigin(),
-	database: drizzleAdapter(database, {
+	baseURL: getAppOrigin(),
+	database: drizzleAdapter(db, {
 		provider: 'pg',
 		schema,
 	}),
@@ -18,12 +18,12 @@ export const auth = betterAuth({
 		github: {
 			clientId: env.GITHUB_CLIENT_ID!,
 			clientSecret: env.GITHUB_CLIENT_SECRET!,
-			redirectURI: `${getOrigin()}/api/auth/callback/github`,
+			redirectURI: `${getAppOrigin()}/api/auth/callback/github`,
 		},
 		discord: {
 			clientId: env.DISCORD_CLIENT_ID!,
 			clientSecret: env.DISCORD_CLIENT_SECRET!,
-			redirectURI: `${getOrigin()}/api/auth/callback/discord`,
+			redirectURI: `${getAppOrigin()}/api/auth/callback/discord`,
 		},
 	} satisfies Record<SupportedOAuthProvider['id'], unknown>,
 	plugins: [sveltekitCookies(getRequestEvent)],
