+++ sites/plus.skeleton.dev/src/lib/server/database/db.ts
@@ -1,9 +1,9 @@
 import { env } from '$env/dynamic/private';
 import { drizzle } from 'drizzle-orm/postgres-js';
 import postgres from 'postgres';
-import * as schema from '$lib/server/database/schema/schema';
+import * as schema from '$lib/server/database/schema';
 
-export const database = drizzle({
+export const db = drizzle({
 	client: postgres(env.DATABASE_URL!, {
 		prepare: false,
 	}),
