+++ sites/plus.skeleton.dev/src/lib/server/database/schema.ts
@@ -1,0 +1,1 @@
+export * from '$lib/server/database/schemas/auth-schema';
