--- sites/plus.skeleton.dev/src/lib/server/database/schema/schema.ts
@@ -1,1 +1,0 @@
-export * from './auth-schema';
